add core locations data source
