+++ anxcloud/provider.go
@@ -30,10 +30,11 @@ func Provider() *schema.Provider {
 			"anxcloud_tag":            resourceTag(),
 		},
 		DataSourcesMap: map[string]*schema.Resource{
-			"anxcloud_disk_type":    dataSourceDiskType(),
-			"anxcloud_template":     dataSourceTemplate(),
-			"anxcloud_ip_addresses": dataSourceIPAddresses(),
-			"anxcloud_nic_types":    dataSourceNICTypes(),
+			"anxcloud_disk_type":      dataSourceDiskType(),
+			"anxcloud_template":       dataSourceTemplate(),
+			"anxcloud_ip_addresses":   dataSourceIPAddresses(),
+			"anxcloud_nic_types":      dataSourceNICTypes(),
+			"anxcloud_core_locations": dataSourceCoreLocations(),
 		},
 		ConfigureContextFunc: providerConfigure,
 	}
